Switch to using crossterm instead of termion
Switch to using crossterm for terminal control instead of using termion.
The main reason for this switch is that I couldn't get a custom panic
hook to work when using termion. I was however able to get it to work
with crossterm. The problem was that the panic hook output was not
visible and so I wanted to try to make it visible.

Crossterm does seem to have a somewhat nicer API as well.

One note is that crossterm has a `queue` macro which improves
readability, but it was producing a warning each invocation :(