Use git rev-parse for repo root
In bash scripts, use `git rev-parse --show-toplevel` to determing the
root directory of the repository instead of determining it relative to
the directory of the bash script. The reason for doing this is that on
Mac Os X the `readlink` command does not have the `-f` flag which was
being used.