Rewrite insh to be component based
Just about compleletly rewrite insh to be component based. What I'm
going for here is something that is like React (although I fall pretty
short). The idea is we should be able to write re-usable components.
Components handle events and optionally emit effects. In general, the
event and effect handling could really use some restructuring.
Especially the fairly common case of delegating in a subcomponent to
handle events. Anyways, I still think this is an improvement but
my lack of experience with Rust really shines through with some of the
inelegance.

NOTE: This rewrite does not (yet) include in the searcher. The rewrite
has taken a while and so I'm going to commit this anyways with the
idea that it will put pressure on myself to rewrite the searcher.