Fix flickering
Fix the flickering of text. Flickering was being caused by clearing the
terminal on each render. Instead, just draw the fabric without clearing
the screen.

There were a number of places where the fabric (or yarns of the fabric)
were not being created with the correct size so fix those too. I had
been thinking that we could use fuzzing to check that yarns and fabrics
are always the right size, but alternatively, it might make sense to
change the render method to have a fabric as a parameter and change the
fabric parameter have checks against being incorrectly sized.