Add searcher history and suggestion
Add storage of searches to `~/.insh/data.yaml` and add auto
complete suggestions to the phrase component. Right now the auto
complete for the searcher just suggests the most recent search that
starts with the same string that has currently been inputted. The
`<Tab>` key can be used to fill in the suggestion.