Fix tabs in the searcher
Fix the displaying of tabs in the line contents of the searcher. They
were seriously messing up the output because they were being printed
directly. Instead, replace them with spaces. The number of spaces is
configurable via the `general.tab_width` setting. Note that it isn't
as simple as a find and replace because tabbing actually aligns the
next input with a column that is a multiple of the tab width.