Add bell
Add the making of the bell sound (via the ASCII bell 0x7).

For now, we only make the bell sound when an invalid command is entered,
not one a valid idempotent one is entered. For example if in the
browser, and the last entry is selected and `j` is entered, then no bell
sound is made.

The bell sound can be turned off via the configuration setting
`general.bell`.